refactor(work): move the submit CLI adapter into the Work service transports tree
Relocate pkg/transports/cli/submit to
pkg/services/work/transports/cli/submit. Its only non-test service import is
pkg/services/work.

Pure relocation: all 17 source files move unchanged and the 11 external
importers are updated to the new import path.

One guard inventory needed the new path to keep covering the same package:
prohibitedFunctionalTransportCompositionOperations in
cmd/pkgboundarycheck/test_behavior_boundary.go keys on the exact submit import
path, so cliSubmitImportPath is repointed. Without that edit the guard would
have silently stopped prohibiting submit.NewSubmit construction from
functional tests. The rule and its finding output are otherwise unchanged.
