+++ pkg/services/work/transports/cli/submit/batch_input.go
