+++ pkg/services/work/transports/cli/submit/batch_json_test.go
