+++ pkg/services/work/transports/cli/submit/error_output.go
