+++ pkg/services/work/transports/cli/submit/output_required_test.go
